Make System.Private.CoreLib compilable on Windows
